+++ testproject/Assets/Tests/Runtime/MultiClientConnectionApproval.cs
@@ -66,6 +66,10 @@ namespace TestProject.RuntimeTests
             m_PlayerPrefab = new GameObject("Player");
             var networkObject = m_PlayerPrefab.AddComponent<NetworkObject>();
 
+            // Prefabs should always be owned by the server
+            // This assures that if a client is shutdown it will not destroy the prefab
+            networkObject.NetworkManagerOwner = server;
+
             // Make it a prefab
             NetcodeIntegrationTestHelpers.MakeNetworkObjectTestPrefab(networkObject);
 
@@ -75,6 +79,9 @@ namespace TestProject.RuntimeTests
                 // Create a default player GameObject to use
                 m_PlayerPrefabOverride = new GameObject("PlayerPrefabOverride");
                 var networkObjectOverride = m_PlayerPrefabOverride.AddComponent<NetworkObject>();
+                // Prefabs should always be owned by the server
+                // This assures that if a client is shutdown it will not destroy the prefab
+                networkObjectOverride.NetworkManagerOwner = server;
                 NetcodeIntegrationTestHelpers.MakeNetworkObjectTestPrefab(networkObjectOverride);
                 m_PrefabOverrideGlobalObjectIdHash = networkObjectOverride.GlobalObjectIdHash;
 
@@ -115,7 +122,6 @@ namespace TestProject.RuntimeTests
                     client.NetworkConfig.ConnectionData = Encoding.ASCII.GetBytes(m_ConnectionToken);
                     clientsAdjustedList.Add(client);
                 }
-
             }
 
             // Start the instances
@@ -148,7 +154,11 @@ namespace TestProject.RuntimeTests
 
             foreach (var client in clients)
             {
-                client.Shutdown();
+                // If a client failed, then it will already be shutdown
+                if (client.IsListening)
+                {
+                    client.Shutdown();
+                }
             }
 
             server.ConnectionApprovalCallback -= ConnectionApprovalCallback;
@@ -227,7 +237,12 @@ namespace TestProject.RuntimeTests
             // [Host-Side] Check to make sure all clients are connected
             yield return NetcodeIntegrationTestHelpers.WaitForClientsConnectedToServer(server, clients.Length + 1, null, 512);
 
+
             Assert.AreEqual(3, m_ClientConnectedInvocations);
+            var timeoutHelper = new TimeoutHelper(2);
+
+            yield return NetcodeIntegrationTest.WaitForConditionOrTimeOut(() => m_ServerClientConnectedInvocations == 4, timeoutHelper);
+            Assert.False(timeoutHelper.TimedOut, $"Timed out waiting for server client connections to reach a count of 4 but only has {m_ServerClientConnectedInvocations}!");
             Assert.AreEqual(4, m_ServerClientConnectedInvocations);
         }
 
@@ -242,7 +257,7 @@ namespace TestProject.RuntimeTests
         }
 
 
-        private int m_ServerClientDisconnectedInvocations;
+        private int m_ClientDisconnectedInvocations;
 
         /// <summary>
         /// Tests that clients are disconnected when their ConnectionApproval setting is mismatched with the host-server
@@ -252,36 +267,34 @@ namespace TestProject.RuntimeTests
         [UnityTest]
         public IEnumerator ConnectionApprovalMismatchTest([Values(true, false)] bool enableSceneManagement, [Values(true, false)] bool connectionApproval)
         {
-            m_ServerClientDisconnectedInvocations = 0;
+            m_ClientDisconnectedInvocations = 0;
 
             // Create Host and (numClients) clients
             Assert.True(NetcodeIntegrationTestHelpers.Create(3, out NetworkManager server, out NetworkManager[] clients));
 
             server.NetworkConfig.EnableSceneManagement = enableSceneManagement;
-            server.OnClientDisconnectCallback += Server_OnClientDisconnectedCallback;
             server.NetworkConfig.ConnectionApproval = connectionApproval;
-
             foreach (var client in clients)
             {
                 client.NetworkConfig.EnableSceneManagement = enableSceneManagement;
                 client.NetworkConfig.ConnectionApproval = !connectionApproval;
+                client.OnClientDisconnectCallback += Client_OnClientDisconnectedCallback; //Server notifies client (not vice versa)
             }
-
             // Start the instances
             if (!NetcodeIntegrationTestHelpers.Start(true, server, clients))
             {
                 Assert.Fail("Failed to start instances");
             }
 
-            var nextFrameNumber = Time.frameCount + 5;
-            yield return new WaitUntil(() => Time.frameCount >= nextFrameNumber);
-
-            Assert.AreEqual(3, m_ServerClientDisconnectedInvocations);
+            var timeoutHelper = new TimeoutHelper();
+            yield return NetcodeIntegrationTest.WaitForConditionOrTimeOut(() => m_ClientDisconnectedInvocations == 3);
+            Assert.False(timeoutHelper.TimedOut, "Timed out waiting for clients to be disconnected!");
+            Assert.AreEqual(3, m_ClientDisconnectedInvocations);
         }
 
-        private void Server_OnClientDisconnectedCallback(ulong clientId)
+        private void Client_OnClientDisconnectedCallback(ulong clientId)
         {
-            m_ServerClientDisconnectedInvocations++;
+            m_ClientDisconnectedInvocations++;
         }